Minor changes to code
- Remove "0" and "1" strings from code and define them at the start of
the Huffman class.
- Removed un-necessary cast to 'int' for bitstring[y].

+++ DataCompression/HuffmanEncoder.cs
@@ -101,9 +101,8 @@ namespace DataCompression
                 for (y = 0; y < bitstring.Length; y++)
                 {
                     // Process the bits into a byte.
-                    //bitVal |= (bitstring[y] == '1' ? 1 : 0);
                     // Subtract 48 from char '1' or '0' to get the integer 1 or 0.
-                    bitVal |= ((int)bitstring[y] - 48);
+                    bitVal |= bitstring[y] - 48;
                     bitVal = bitVal << 1;
                     z++;
